complete category crud

+++ server/routs/categories.route.js
@@ -1,76 +1,64 @@
 const router = require('express').Router()
 const Category = require('../models/Category')
+const Product = require('../models/Product')
 const multer = require('multer');
 const path = require('path');
 const fs = require('fs');
-
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, path.join(__dirname, '..', 'uploads'))
-  },
-  filename: function (req, file, cb) {
-
-    let fileExtention = '';
-    for (let i = file.originalname.length-1; i >= 0; i--) {
-      if (file.originalname[i] === '.'){
-        break
-      }
-      fileExtention += file.originalname[i]
-    }
-
-    cb(null, Date.now() + '.' + fileExtention.split('').reverse().join(''))
-  }
-})
-
-const upload = multer({ storage: storage })
 
 router.get('/', async (req, res) => {
   const page = parseInt(req.query.page)
   const skip = (page-1)*10
   const { search } = req.query
   try {
-    const categories = await Category.find({name: {$regex: search}}).limit(10).skip(skip);
-    res.status(200).json(categories)
+    const categories = await Category.find({name: {$regex: search}}).limit(8).skip(skip);
+    const count = await Category.find({name: {$regex: search}}).count()
+    res.status(200).json({categories, count})
   } catch (err) {
     console.log(err)
     res.status(400).json('some thing went wrong')
   }
 })
 
-router.post('/', upload.single('image'), async (req, res) => {
-  const {name} = req.body
+router.get('/onecategory/:id', async (req, res) => {
+  const { id } = req.params
   try {
-    if (req.file) {
-      const image = `${req.protocol}://${req.get('host')}/uploads/${req.file.filename}`
-      const newCategory = new Category({name, image})
-      await newCategory.save()
-      res.status(200).json(newCategory)
-    } else {
-      res.status(400).json('some thing went wrong')
-    }
+    const category = await Category.findById(id)
+    res.status(200).json(category)
+  } catch (err) {
+    console.log(err)
+    res.status(400).json('some thing went wrong')
+  }
+})
+
+router.post('/', async (req, res) => {
+  const {name, image, filters} = req.body
+  if (filters.split(',').length>10) return res.status(400).json('some thing went wrong with filters')
+  try {
+    const newImage = `${req.protocol}://${req.get('host')}/uploads/${image.url}`
+    const newCategory = new Category({name, image: newImage, filters: filters ? filters.split(',') : []})
+    await newCategory.save()
+    res.status(200).json('added successfully')
   } catch(err) {
     console.log(err)
     res.status(400).json('some thing went wrong')
   }
 })
 
-router.put('/:id',  upload.single('image'), async (req, res) => {
+router.put('/:id', async (req, res) => {
   const {id} = req.params
-  const {name} = req.body
+  const {name, image, filters} = req.body
+  if (filters.split(',').length>10) return res.status(400).json('some thing went wrong with filters')
   try {
-    if (req.file) {
-      const image = `${req.protocol}://${req.get('host')}/uploads/${req.file.filename}`
-      const category = await Category.findByIdAndUpdate(id, {name, image})
-      const imageName = category?.image.split('/')[category?.image.split('/').length-1]
+    const imageUrl = image.touched ? `${req.protocol}://${req.get('host')}/uploads/${image.url}` : image.url
+    const category = await Category.findByIdAndUpdate(id, {name, image: imageUrl, filters: filters ? filters.split(',') : []})
+    const imageName = category?.image.split('/')[category?.image.split('/').length-1]
+    if (image.touched) {
       const fileTest = fs.existsSync(path.join(__dirname, '..', 'uploads', imageName))
       if (fileTest) {
         fs.unlinkSync(path.join(__dirname, '..', 'uploads', imageName))
       }
-      res.status(200).json('updated successfully')
-    } else {
-      await Category.updateOne({_id: id}, {name})
-      res.status(200).json('updated successfully')
     }
+    res.status(200).json('updated successfully')
   } catch(err) {
     console.log(err)
     res.status(400).json('some thing went wrong')
@@ -80,13 +68,18 @@ router.put('/:id',  upload.single('image'), async (req, res) => {
 router.delete('/:id', async (req, res) => {
   const {id} = req.params
   try {
-    const category = await Category.findByIdAndDelete(id)
-    const imageName = category?.image.split('/')[category?.image.split('/').length-1]
-    const fileTest = fs.existsSync(path.join(__dirname, '..', 'uploads', imageName))
-      if (fileTest) {
-        fs.unlinkSync(path.join(__dirname, '..', 'uploads', imageName))
-      }
-    res.status(200).json('deleted successfully')
+    const prod = await Product.findOne({category: id})
+    if (!prod?._id) {
+      const category = await Category.findByIdAndDelete(id)
+      const imageName = category?.image.split('/')[category?.image.split('/').length-1]
+      const fileTest = fs.existsSync(path.join(__dirname, '..', 'uploads', imageName))
+        if (fileTest) {
+          fs.unlinkSync(path.join(__dirname, '..', 'uploads', imageName))
+        }
+      res.status(200).json('deleted successfully')
+    } else {
+      res.status(400).json('the category must have no products')
+    }
   } catch (err) {
     console.log(err)
     res.status(400).json('some thing went wrong')
